## 安装onediff及其依赖
```shell
 python3 -m pip install --pre oneflow -f https://oneflow-pro.oss-cn-beijing.aliyuncs.com/branch/community/cu121 && 
git clone https://github.com/siliconflow/onediff.git &&
cd onediff &&
python3 -m pip install -e . &&
cd onediff_diffusers_extensions &&
python3 -m pip install -e .
```

## 加速
- warmup 2 times
- speed:
  - baseline: 2.77it/s
  - onediff: 5.18it/s
  - (5.18 - 2.77) / 2.77 = 0.87 (87% faster)
  
- e2e:
  - baseline: 20.18 s
  - onediff: 11.88 s
  -  (20.18 - 11.88) / 20.18 = 0.41 (41% faster)
 
添加 📷OneDiff Speed Up 如下图：
![workflow (11)](https://github.com/ZHO-ZHO-ZHO/ComfyUI-InstantID/assets/109639975/f5bc8c9c-6bb2-4954-8d00-c6c512dddb53)